ProgrammersApi/requirements.txt
asgiref==3.3.0
Django==3.0.8
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.4.1
